+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Version 7.1
 
+### Version 7.1.2.3 - 2021-02-03
+
+#### Added
+-   Class `GPUdb.GPUdbVersion` that represents the Kinetica server's
+    version (the one the API is connected to).
+-   Method `GPUdb.getServerVersion()`.
+
+
+#### Changed
+-   Added support for multi-head key lookup for replicated tables.
+
+
+
 ### Version 7.1.2.2 - 2021-01-29
 
 #### Security
@@ -132,6 +145,19 @@
 
 ## Version 7.0
 
+### Version 7.0.20.4 - 2021-02-02
+
+#### Added
+-   Class `GPUdb.GPUdbVersion` that represents the Kinetica server's
+    version (the one the API is connected to).
+-   Method `GPUdb.getServerVersion()`.
+
+
+#### Changed
+-   Added support for multi-head key lookup for replicated tables.
+
+
+
 ### Version 7.0.20.3 - 2021-01-29
 
 #### Security
@@ -168,6 +194,25 @@
 
 #### Note
 -   Check CHANGELOG-FUNCTIONS.md for endpoint related changes.
+
+
+### Version 7.0.19.3 - 2021-02-01
+
+#### Added
+-   Class `GPUdb.GPUdbVersion` that represents the Kinetica server's
+    version (the one the API is connected to).
+-   Method `GPUdb.getServerVersion()`.
+
+
+#### Changed
+-   Added support for multi-head key lookup for replicated tables.
+
+
+
+### Version 7.0.19.2 - 2021-01-26
+
+#### Fixed
+-   An issue with BulkInserter flush when retryCount > 0
 
 
 ### Version 7.0.19.1 - 2020-12-23